remove old header and footer and update links
