--- app/components/footer-new/index.njk
@@ -1,33 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Footer short list (new)' %}
-{% from 'components/footer-new/macro.njk' import footerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ footerNew({
-    "links": [
-      {
-        "URL": "#",
-        "label": "Accessibility statement"
-      },
-      {
-        "URL": "#",
-        "label": "Contact us"
-      },
-      {
-        "URL": "#",
-        "label": "Cookies"
-      },
-      {
-        "URL": "#",
-        "label": "Privacy policy"
-      },
-      {
-        "URL": "#",
-        "label": "Terms and conditions"
-      }
-    ]
-  })}}
-
-{% endblock %}
