+++ app/components/footer/index.njk
@@ -1,5 +1,5 @@
 {% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Footer' %}
+{% set title = 'Footer (default)' %}
 {% from 'components/footer/macro.njk' import footer %}
 {% extends 'layout.njk' %}
