--- app/components/header-new/header-logo.njk
@@ -1,14 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-
-{% endblock %}
