--- app/components/header-new/header-org-white-nav.njk
@@ -1,42 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header organisational with white header and navigation (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-{{ headerNew({
-    "showNav": "true",
-    "showSearch": "true",
-    "classes": "beta-nhsuk-header--white beta-nhsuk-header--white-nav",
-    "organisation": {
-      "name": "Anytown Anyplace",
-      "split": "Anywhere",
-      "descriptor": "NHS Foundation Trust"
-    },
-    "primaryLinks": [
-      {
-        "url"  : "#",
-        "label" : "Your hospital visit"
-      },
-      {
-        'url' : '#',
-        'label' : 'Wards and departments'
-      },
-      {
-        'url'  : '#',
-        'label' : 'Conditions and treatments'
-      },
-      {
-        'url'  : '#',
-        'label' : 'Our people'
-      },
-      {
-        'url' : '#',
-        'label' : 'Our research'
-      }
-    ]
-  })
-}}
-
-{% endblock %}
