--- app/components/header-new/header-service-name-with-nav.njk
@@ -1,39 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header with a service name, search and navigation (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-    "service": {
-      "name": "Digital service manual"
-    },
-    "showNav": "true",
-    "showSearch": "true",
-    "primaryLinks": [
-        {
-          "url"  : "#",
-          "label" : "NHS service standard"
-        },
-        {
-          'url' : '#',
-          'label' : 'Design system'
-        },
-        {
-          'url'  : '#',
-          'label' : 'Content style guide'
-        },
-        {
-          'url'  : '#',
-          'label' : 'Accessibility'
-        },
-        {
-          'url' : '#',
-          'label' : 'Community and contribution'
-        }
-      ]
-    })
-  }}
-
-{% endblock %}
