--- app/components/header-new/header-service-name.njk
@@ -1,15 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header with service name (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-      "service": {
-        "name": "Prototype kit"
-      }
-    })
-  }}
-
-{% endblock %}
