--- app/components/header-new/header-transactional-service-name.njk
@@ -1,17 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header transactional with service name (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-      "transactionalService": {
-        "name": "Find your NHS number"
-      },
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-
-{% endblock %}
