--- app/components/header-new/header-transactional.njk
@@ -1,15 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header transactional (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-      "transactional": "true",
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-
-{% endblock %}
