--- app/components/header-new/index.njk
@@ -1,40 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Default header (new)' %}
-{% from 'components/header-new/macro.njk' import headerNew %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ headerNew({
-      "showNav": "true",
-      "showSearch": "true",
-      "primaryLinks": [
-        {
-          "url"  : "https://www.nhs.uk/conditions",
-          "label" : "Health A-Z"
-        },
-        {
-          'url' : 'https://www.nhs.uk/nhs-services/',
-          'label' : 'NHS services'
-        },
-        {
-          'url' : 'https://www.nhs.uk/live-well/',
-          'label' : 'Live Well'
-        },
-        {
-          'url' : 'https://www.nhs.uk/mental-health/',
-          'label' : 'Mental health'
-        },
-        {
-          'url'  : 'https://www.nhs.uk/conditions/social-care-and-support/',
-          'label' : 'Care and support'
-        },
-        {
-          'url'  : 'https://www.nhs.uk/pregnancy/',
-          'label' : 'Pregnancy'
-        }
-      ]
-    })
-  }}
-
-{% endblock %}
