+++ app/components/header/header-org-white-nav.njk
@@ -8,7 +8,7 @@
 {{ header({
     "showNav": "true",
     "showSearch": "true",
-    "classes": "nhsuk-header--white nhsuk-header--white-nav",
+    "classes": "beta-nhsuk-header--white beta-nhsuk-header--white-nav",
     "organisation": {
       "name": "Anytown Anyplace",
       "split": "Anywhere",
