+++ app/components/header/header-org-white.njk
@@ -8,7 +8,7 @@
 {{ header({
     "showNav": "true",
     "showSearch": "true",
-    "classes": "nhsuk-header--white",
+    "classes": "beta-nhsuk-header--white",
     "organisation": {
       "name": "Anytown Anyplace",
       "split": "Anywhere",
