+++ app/components/header/header-service-name-with-nav.njk
@@ -1,5 +1,5 @@
 {% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header transactional with service name' %}
+{% set title = 'Header with a service name, search and navigation' %}
 {% from 'components/header/macro.njk' import header %}
 {% extends 'layout.njk' %}
 
@@ -14,7 +14,7 @@
     "primaryLinks": [
         {
           "url"  : "#",
-          "label" : "NHS service standard"
+          "label" : "Standards and technology"
         },
         {
           'url' : '#',
@@ -27,6 +27,10 @@
         {
           'url'  : '#',
           'label' : 'Accessibility'
+        },
+        {
+          'url'  : '#',
+          'label' : 'Design principles'
         },
         {
           'url' : '#',
