+++ app/components/header/header-transactional-service-name.njk
@@ -7,7 +7,7 @@
 
   {{ header({
       "transactionalService": {
-        "name": "Register with a GP"
+        "name": "Find your NHS number"
       },
       "showNav": "false",
       "showSearch": "false"
