--- app/components/header/header-transactional.njk
@@ -1,15 +1,0 @@
-{% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header transactional' %}
-{% from 'components/header/macro.njk' import header %}
-{% extends 'layout.njk' %}
-
-{% block body %}
-
-  {{ header({
-      "transactional": "true",
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-
-{% endblock %}
