+++ app/components/header/index.njk
@@ -1,5 +1,5 @@
 {% set html_style = 'background-color: #f0f4f5;' %}
-{% set title = 'Header' %}
+{% set title = 'Header (default)' %}
 {% from 'components/header/macro.njk' import header %}
 {% extends 'layout.njk' %}
