+++ app/pages/examples.njk
@@ -68,31 +68,18 @@
     <li><a href="../components/fieldset/index.html">Fieldset</a></li>
     <li><a href="../components/fieldset/page-heading.html">Fieldset as page heading</a></li>
     <li><a href="../components/fieldset/with-inputs.html">Fieldset with inputs</a></li>
-    <li><a href="../components/footer/index.html">Footer</a></li>
-    <li><a href="../components/footer-new/index.html">Footer short list (new)</a></li>
-    <li><a href="../components/footer-new/footer-long-list.html">Footer long list (new)</a></li>
-    <li><a href="../components/header/index.html">Header</a></li>
+    <li><a href="../components/footer/index.html">Footer (default)</a></li>
+    <li><a href="../components/footer/footer-in-columns.html">Footer (columns)</a></li>
+    <li><a href="../components/header/index.html">Header (default)</a></li>
     <li><a href="../components/header/header-navigation.html">Header with navigation</a></li>
     <li><a href="../components/header/header-search.html">Header with search</a></li>
     <li><a href="../components/header/header-logo.html">Header with logo only</a></li>
     <li><a href="../components/header/header-service-name.html">Header with a service name</a></li>
     <li><a href="../components/header/header-service-name-with-nav.html">Header with a service name, search and navigation</a></li>
-    <li><a href="../components/header/header-transactional.html">Header transactional</a></li>
     <li><a href="../components/header/header-transactional-service-name.html">Header transactional with service name</a></li>
     <li><a href="../components/header/header-org.html">Header organisational</a></li>
     <li><a href="../components/header/header-org-white.html">Header organisational with white header</a></li>
     <li><a href="../components/header/header-org-white-nav.html">Header organisational with white header and navigation</a></li>
-    <li><a href="../components/header-new/index.html">Default header (new)</a></li>
-    <li><a href="../components/header-new/header-navigation.html">Header with navigation (new)</a></li>
-    <li><a href="../components/header-new/header-search.html">Header with search (new)</a></li>
-    <li><a href="../components/header-new/header-logo.html">Header with logo only (new)</a></li>
-    <li><a href="../components/header-new/header-service-name.html">Header with a service name (new)</a></li>
-    <li><a href="../components/header-new/header-service-name-with-nav.html">Header with a service name, search and navigation (new)</a></li>
-    <li><a href="../components/header-new/header-transactional.html">Header transactional (new)</a></li>
-    <li><a href="../components/header-new/header-transactional-service-name.html">Header transactional with service name (new)</a></li>
-    <li><a href="../components/header-new/header-org.html">Header organisational (new)</a></li>
-    <li><a href="../components/header-new/header-org-white.html">Header organisational with white header (new)</a></li>
-    <li><a href="../components/header-new/header-org-white-nav.html">Header organisational with white header and navigation (new)</a></li>
     <li><a href="../components/hero/index.html">Hero</a></li>
     <li><a href="../components/hero/hero-image.html">Hero with image</a></li>
     <li><a href="../components/hero/hero-image-content.html">Hero with image with content</a></li>
