--- packages/components/footer-new/_footer-new.scss
@@ -1,63 +1,0 @@
-/* ==========================================================================
-   COMPONENTS / #FOOTER-NEW
-   ========================================================================== */
-
-.beta-nhsuk-footer {
-
-  background-color: $color_nhsuk-grey-4;
-  // border-top: nhsuk-spacing(1) solid $color_nhsuk-blue;
-  margin-bottom: 15px;
-
-  @include mq($from: desktop) {
-    display: flex;
-  }
-}
-
-
-.beta-nhsuk-footer__list {
-  margin-bottom: 25px;
-
-  &:last-child {
-    margin-bottom: 15px;
-  }
-
-  @include mq($from: desktop) {
-    padding-right: 40px;
-
-    &:last-child {
-      padding-right: 0;
-    }
-  }
-}
-
-.beta-nhsuk-footer__list-policies {
-  @include mq($until: desktop) {
-    border-top:1px solid #f0f4f5;
-    padding-top: 35px
-    }
-}
-
-.beta-nhsuk-footer__list-item {
-  @include mq($from: desktop) {
-    float: none;
-    margin-right: 0;
-  }
-}
-
-$beta-link-aaa: rgb(0, 48, 135);
-$beta-copyright-aaa: rgb(35, 31, 32);
-
-.beta-nhsuk-footer__list .beta-nhsuk-footer__list-item-link {
-  color: $beta-link-aaa;
-
-  &:hover {
-    color: $nhsuk-link-hover-color;
-    text-decoration: none;
-  }
-}
-
-.beta-nhsuk-footer__copyright {
-  @include nhsuk-font(16);
-  color: $beta-copyright-aaa;
-  margin-bottom: 0;
-}
