--- packages/components/footer-new/macro.njk
@@ -1,3 +1,0 @@
-{% macro footerNew(params) %}
-  {%- include './template.njk' -%}
-{% endmacro %}
