+++ packages/components/header/header.js
@@ -1,7 +1,5 @@
 import MenuToggle from './menuToggle'
-import SearchToggle from './searchToggle'
 
 export default () => {
   MenuToggle()
-  SearchToggle()
 }
