+++ packages/components/headerOld/header.js
@@ -1,5 +1,7 @@
 import MenuToggle from './menuToggle'
+import SearchToggle from './searchToggle'
 
 export default () => {
   MenuToggle()
+  SearchToggle()
 }
