+++ packages/components/headerOld/header.test.js
@@ -1,9 +1,9 @@
-import HeaderNew from './header'
+import Header from './header'
 
 describe('NHS.UK header()', () => {
   describe('Does not throw an error', () => {
     it('if no document exists', () => {
-      HeaderNew()
+      Header()
     })
   })
 })
