+++ packages/components/headerOld/macro.njk
@@ -1,3 +1,3 @@
-{% macro headerNew(params) %}
+{% macro header(params) %}
   {%- include './template.njk' -%}
 {% endmacro %}
