+++ packages/nhsuk.js
@@ -5,7 +5,6 @@ import Checkboxes from './components/checkboxes/checkboxes'
 import Details from './components/details/details'
 import ErrorSummary from './components/error-summary/error-summary'
 import Header from './components/header/header'
-import HeaderNew from './components/header-new/header'
 import Radios from './components/radios/radios'
 import SkipLink from './components/skip-link/skip-link'
 import Tabs from './components/tabs/tabs'
@@ -20,7 +19,6 @@ document.addEventListener('DOMContentLoaded', () => {
   Details()
   ErrorSummary()
   Header()
-  HeaderNew()
   Radios()
   SkipLink()
   Tabs()
