+++ packages/nhsuk.scss
@@ -15,9 +15,7 @@
 @import "components/error-summary/error-summary";
 @import "components/fieldset/fieldset";
 @import "components/footer/footer";
-@import "components/footer-new/footer-new";
 @import "components/header/header";
-@import "components/header-new/header-new";
 @import "components/hero/hero";
 @import "components/hint/hint";
 @import "components/images/images";
